chore: move swappers to package (pt 1) (#5792)
* chore: add package structure

* chore: update yarn.lock with new package

* chore: update yarn.lock with new package

* chore: iron out importing

* chore: move out some types as Poc

* chore: address review comments

* chore: revert star export removal
