+++ packages/swapper/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@shapeshiftoss/swapper",
+  "packageManager": "yarn@3.5.0",
+  "repository": "https://github.com/shapeshift/web",
+  "license": "MIT",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "build": "yarn clean && yarn run -T tsc --build",
+    "clean": "rm -rf dist tsconfig.tsbuildinfo",
+    "dev": "yarn run -T tsc --build --watch"
+  },
+  "dependencies": {
+    "@shapeshiftoss/caip": "workspace:^"
+  }
+}
