+++ packages/swapper/tsconfig.json
@@ -11,4 +11,12 @@
   "exclude": [
     "dist"
   ],
+  "references": [
+    {
+      "path": "../types",
+    },
+    {
+      "path": "../caip",
+    }
+  ]
 }
