+++ src/components/MultiHopTrade/components/TradeInput/components/TradeQuotes/getQuoteErrorTranslation.ts
@@ -1,10 +1,9 @@
 import type { AssetId } from '@shapeshiftoss/caip'
 import { bnOrZero } from '@shapeshiftoss/chain-adapters'
+import { type SwapErrorRight, SwapErrorType } from '@shapeshiftoss/swapper'
 import type { InterpolationOptions } from 'node-polyglot'
 import type { Asset } from 'lib/asset-service'
 import { baseUnitToHuman } from 'lib/bignumber/bignumber'
-import type { SwapErrorRight } from 'lib/swapper/types'
-import { SwapErrorType } from 'lib/swapper/types'
 import type { PartialRecord } from 'lib/utils'
 
 export const quoteStatusTranslation = (
