+++ src/components/MultiHopTrade/utils.ts
@@ -1,10 +1,10 @@
 import type { AssetId, ChainId } from '@shapeshiftoss/caip'
+import type { SwapErrorRight } from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import axios from 'axios'
 import { getConfig } from 'config'
 import { isRune } from 'lib/swapper/swappers/ThorchainSwapper/utils/isRune/isRune'
-import type { SwapErrorRight } from 'lib/swapper/types'
 import { SwapperName } from 'lib/swapper/types'
 import { getInboundAddressDataForChain } from 'lib/utils/thorchain/getInboundAddressDataForChain'
