+++ src/hooks/useErrorToast/useErrorToast.ts
@@ -1,8 +1,8 @@
 import { useToast } from '@chakra-ui/react'
+import { SwapErrorType } from '@shapeshiftoss/swapper'
 import { get, isError } from 'lodash'
 import { useCallback } from 'react'
 import { useTranslate } from 'react-polyglot'
-import { SwapErrorType } from 'lib/swapper/types'
 
 // TODO support more error types (non swapper errors)
 export const ErrorTranslationMap: Record<string, string> = {
