+++ src/lib/swapper/swappers/CowSwapper/getCowSwapTradeQuote/getCowSwapTradeQuote.ts
@@ -1,4 +1,5 @@
 import { fromAssetId } from '@shapeshiftoss/caip'
+import type { SwapErrorRight } from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import type { AxiosError } from 'axios'
@@ -23,7 +24,7 @@ import {
   isNativeEvmAsset,
   normalizeIntegerAmount,
 } from 'lib/swapper/swappers/utils/helpers/helpers'
-import type { GetTradeQuoteInput, SwapErrorRight, TradeQuote } from 'lib/swapper/types'
+import type { GetTradeQuoteInput, TradeQuote } from 'lib/swapper/types'
 import { SwapperName } from 'lib/swapper/types'
 import { createTradeAmountTooSmallErr } from 'lib/swapper/utils'
