+++ src/lib/swapper/swappers/OneInchSwapper/endpoints.ts
@@ -1,5 +1,6 @@
 import { fromChainId } from '@shapeshiftoss/caip'
 import type { EvmChainId } from '@shapeshiftoss/chain-adapters'
+import type { SwapErrorRight } from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads/build'
 import { v4 as uuid } from 'uuid'
 import type {
@@ -7,7 +8,6 @@ import type {
   GetEvmTradeQuoteInput,
   GetTradeQuoteInput,
   GetUnsignedEvmTransactionArgs,
-  SwapErrorRight,
   SwapperApi,
   TradeQuote,
 } from 'lib/swapper/types'
