+++ src/lib/swapper/swappers/ThorchainSwapper/cosmossdk/getCosmosTxData.ts
@@ -2,14 +2,15 @@ import type { ChainId } from '@shapeshiftoss/caip'
 import { cosmosAssetId } from '@shapeshiftoss/caip'
 import type { CosmosSdkBaseAdapter, thorchain } from '@shapeshiftoss/chain-adapters'
 import type { CosmosSignTx, ThorchainSignTx } from '@shapeshiftoss/hdwallet-core'
+import type { SwapErrorRight } from '@shapeshiftoss/swapper'
+import { SwapErrorType } from '@shapeshiftoss/swapper'
 import { KnownChainIds } from '@shapeshiftoss/types'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import { getConfig } from 'config'
 import type { Asset } from 'lib/asset-service'
 import type { ThorCosmosSdkSupportedChainId } from 'lib/swapper/swappers/ThorchainSwapper/types'
-import type { CosmosSdkFeeData, SwapErrorRight, TradeQuote } from 'lib/swapper/types'
-import { SwapErrorType } from 'lib/swapper/types'
+import type { CosmosSdkFeeData, TradeQuote } from 'lib/swapper/types'
 import { makeSwapErrorRight } from 'lib/swapper/utils'
 import { getInboundAddressDataForChain } from 'lib/utils/thorchain/getInboundAddressDataForChain'
