+++ src/lib/swapper/swappers/ThorchainSwapper/getThorTradeQuote/getTradeQuote.ts
@@ -1,11 +1,12 @@
 import { fromAssetId } from '@shapeshiftoss/caip'
+import type { SwapErrorRight } from '@shapeshiftoss/swapper'
+import { SwapErrorType } from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads'
 import { Err } from '@sniptt/monads'
 import { getConfig } from 'config'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
 import { bn } from 'lib/bignumber/bignumber'
-import type { GetTradeQuoteInput, SwapErrorRight, TradeQuote } from 'lib/swapper/types'
-import { SwapErrorType } from 'lib/swapper/types'
+import type { GetTradeQuoteInput, TradeQuote } from 'lib/swapper/types'
 import { makeSwapErrorRight } from 'lib/swapper/utils'
 import { assertUnreachable } from 'lib/utils'
 import type { AssetsById } from 'state/slices/assetsSlice/assetsSlice'
