+++ src/lib/swapper/swappers/ThorchainSwapper/utils/getLongtailQuote.ts
@@ -1,4 +1,5 @@
 import type { EvmChainId } from '@shapeshiftoss/chain-adapters'
+import { type SwapErrorRight, SwapErrorType } from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import { computePoolAddress, FeeAmount } from '@uniswap/v3-sdk'
@@ -6,7 +7,7 @@ import assert from 'assert'
 import type { Address } from 'viem'
 import { getContract } from 'viem'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
-import { type GetTradeQuoteInput, type SwapErrorRight, SwapErrorType } from 'lib/swapper/types'
+import { type GetTradeQuoteInput } from 'lib/swapper/types'
 import { makeSwapErrorRight } from 'lib/swapper/utils'
 import { viemClientByChainId } from 'lib/viem-client'
 import type { AssetsById } from 'state/slices/assetsSlice/assetsSlice'
