+++ src/lib/swapper/swappers/ZrxSwapper/endpoints.ts
@@ -1,11 +1,11 @@
 import { fromChainId } from '@shapeshiftoss/caip'
+import type { SwapErrorRight } from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads/build'
 import type {
   EvmTransactionRequest,
   GetEvmTradeQuoteInput,
   GetTradeQuoteInput,
   GetUnsignedEvmTransactionArgs,
-  SwapErrorRight,
   SwapperApi,
   TradeQuote,
 } from 'lib/swapper/types'
