+++ src/state/apis/swappers/types.ts
@@ -1,5 +1,6 @@
+import type { SwapErrorRight } from '@shapeshiftoss/swapper'
 import type { Result } from '@sniptt/monads'
-import type { GetTradeQuoteInput, SwapErrorRight, SwapperName, TradeQuote } from 'lib/swapper/types'
+import type { GetTradeQuoteInput, SwapperName, TradeQuote } from 'lib/swapper/types'
 import type { ReduxState } from 'state/reducer'
 
 export type QuoteHelperType = (
