+++ src/state/slices/tradeQuoteSlice/selectors.ts
@@ -1,11 +1,12 @@
 import { createSelector } from '@reduxjs/toolkit'
 import type { AssetId } from '@shapeshiftoss/caip'
+import type { SwapErrorRight } from '@shapeshiftoss/swapper'
 import { getDefaultSlippageDecimalPercentageForSwapper } from 'constants/constants'
 import type { Selector } from 'reselect'
 import type { Asset } from 'lib/asset-service'
 import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
-import type { ProtocolFee, SwapErrorRight, TradeQuote } from 'lib/swapper/types'
+import type { ProtocolFee, TradeQuote } from 'lib/swapper/types'
 import { SwapperName } from 'lib/swapper/types'
 import type { ApiQuote } from 'state/apis/swappers'
 import { selectSwappersApiTradeQuotes } from 'state/apis/swappers/selectors'
