+++ tsconfig.packages.json
@@ -9,10 +9,23 @@
   "include": [],
   "exclude": [],
   "references": [
-    { "path": "./packages/caip" },
-    { "path": "./packages/chain-adapters" },
-    { "path": "./packages/errors" },
-    { "path": "./packages/types" },
-    { "path": "./packages/unchained-client" },
+    {
+      "path": "./packages/caip"
+    },
+    {
+      "path": "./packages/chain-adapters"
+    },
+    {
+      "path": "./packages/errors"
+    },
+    {
+      "path": "./packages/types"
+    },
+    {
+      "path": "./packages/unchained-client"
+    },
+    {
+      "path": "./packages/swapper"
+    },
   ]
 }
